Fix state rollback upon exception
Previously, after update/handle raised an exception, the state wouldn't be modified. But this allowed the app to get into an inconsistent state - if an `update` had an effect, but `handle` raised, the app would look like the action succeeded even though the effect was never executed. For example, If the user set a filter to an illegal frequency, `handle` would fail, but the ui would still display that frequency, looking like nothing happened.

This change fixes that – upon an exception, the state from the previous state is restored.